requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
pypdf
sentence-transformers
python-dotenv
tiktoken
requests
